WIP 2: possible move import to a module, but not possible make it reusable, since imports have to be literals
